Update CombineMenu project for v3
- update to v3.0.0 SDK
- update models and upsert calls to deprecate Codable support
- create separate view model files and folders for organization
- add Info.plist with UISceneConfigurations key to fix Xcode 14.2 console warnings
