+++ combine-menu/CombineMenu.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -5,8 +5,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/getditto/DittoSwiftPackage",
       "state" : {
-        "revision" : "a09b6e3fcdb14efae847db85aa4ea69d74555aeb",
-        "version" : "2.0.0-alpha1"
+        "revision" : "0f41078a6a7441cc56c6be952fa6bcd85b4e0199",
+        "version" : "3.0.0"
       }
     }
   ],
