+++ combine-menu/CombineMenu/Data/Product.swift
@@ -5,18 +5,19 @@
 //  Created by Maximilian Alexander on 3/3/22.
 //
 
-import Foundation
 import DittoSwift
+import Foundation
 
-struct Product: Codable {
-    var _id: String
+struct Product: Identifiable, Equatable, Hashable {
+    var id: String
     var name: String
     var detail: String
     var categoryId: String
-}
-
-extension Product: Identifiable {
-    var id: String {
-        return self._id
+    
+    init(document: DittoDocument) {
+        self.id = document["_id"].stringValue
+        self.name = document["name"].stringValue
+        self.detail = document["detail"].stringValue
+        self.categoryId = document["categoryId"].stringValue
     }
 }
